pyproject.toml
[tool.poetry]
name = "podcast-to-reels"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "podcast_to_reels"}]

[tool.poetry.dependencies]
python = "^3.11"
yt-dlp = "^2023.12.30"
openai = "^1.3.7" # Using a recent version, adjust as needed
fasttext = "^0.9.2"
langdetect = "^1.0.9"
requests = "^2.31.0"
moviepy = "^1.0.3"
Pillow = "^10.0.1" # For image manipulation, used in example, good companion for moviepy
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
ruff = "^0.1.9" # Or your preferred version

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
